fix: organize users controller files on same directory
