+++ src/http/controllers/users/routes.ts
@@ -1,8 +1,8 @@
+import { verifyJwt } from '@/http/middlewares/verify-jwt'
 import { FastifyInstance } from 'fastify'
-import { authenticate } from './controllers/authenticate'
-import { profile } from './controllers/profile'
-import { register } from './controllers/register'
-import { verifyJwt } from './middlewares/verify-jwt'
+import { authenticate } from './authenticate'
+import { profile } from './profile'
+import { register } from './register'
 
 export async function appRoutes(app: FastifyInstance) {
   app.post('/users', register)
